Plan 201: rename internal/testutil to internal/testsymlink
The package name was the canonical util/helpers smell. The
package held one helper, symlink.go, whose job is to skip tests
when the host cannot create symbolic links. Renaming to
testsymlink names the question its single file answers and
deters future grab-bag growth.

Five test files updated to the new import path; the audit log
entry moved from "tax" to a new "resolved by plan/201" section.
