+++ internal/githooks/githooks_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 
 	"github.com/jeduden/mdsmith/internal/config"
-	"github.com/jeduden/mdsmith/internal/testutil"
+	"github.com/jeduden/mdsmith/internal/testsymlink"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
@@ -256,7 +256,7 @@ func TestDiscoverFiles_IgnoresDirectiveMentionsInProse(t *testing.T) {
 }
 
 func TestDiscoverFiles_SkipsSymlinks(t *testing.T) {
-	testutil.SkipIfSymlinkUnsupported(t)
+	testsymlink.SkipIfSymlinkUnsupported(t)
 	dir := t.TempDir()
 	// A real file with a directive that should be discovered.
 	require.NoError(t, os.WriteFile(filepath.Join(dir, "real.md"),
